requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
pytest==7.4.3
gunicorn==21.2.0
psycopg2-binary==2.9.9 